TR-1286 Moved loading of templates from page component to context provider
