+++ src/pages/templates/EditAndPreviewPage.js
@@ -1,28 +1,13 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { Grid, Page } from '@sparkpost/matchbox';
 import { RedirectAndAlert } from 'src/components/globalAlert';
 import Loading from 'src/components/loading';
-import useRouter from 'src/hooks/useRouter';
 import EditPrimaryArea from './components/EditPrimaryArea';
 import EditSection from './components/EditSection';
 import useEditorContext from './hooks/useEditorContext';
 
 const EditAndPreviewPage = () => {
-  const {
-    draft,
-    getDraft,
-    getPublished,
-    hasDraftFailedToLoad,
-    isDraftLoading
-  } = useEditorContext();
-  const { requestParams } = useRouter();
-
-  // hydrate
-  // todo, maybe move to the context provider?
-  useEffect(() => {
-    getDraft(requestParams.id, requestParams.subaccount);
-    getPublished(requestParams.id, requestParams.subaccount);
-  }, [getDraft, getPublished, requestParams.id, requestParams.subaccount]);
+  const { draft, hasDraftFailedToLoad, isDraftLoading } = useEditorContext();
 
   if (hasDraftFailedToLoad) {
     return (
