+++ src/pages/templates/context/tests/EditorContext.test.js
@@ -1,16 +1,24 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { mount, shallow } from 'enzyme';
+import useRouter from 'src/hooks/useRouter';
 import { EditorContextProvider } from '../EditorContext';
 
+jest.mock('src/hooks/useRouter');
 jest.mock('../../hooks/useEditorContent', () => () => ({ content: { text: 'Example' }}));
 
 describe('EditorContext', () => {
   describe('EditorContextProvider', () => {
-    const subject = ({ value = {}} = {}) => shallow(
-      <EditorContextProvider value={value}>
-        <div>Children</div>
-      </EditorContextProvider>
-    );
+    const subject = ({ render = shallow, value = {}} = {}) => {
+      useRouter.mockReturnValue({ requestParams: { id: 'test-template', subaccount: '123' }});
+
+      return render(
+        <EditorContextProvider
+          value={{ getDraft: () => {}, getPublished: () => {}, ...value }}
+        >
+          <div>Hello</div>
+        </EditorContextProvider>
+      );
+    };
 
     it('renders children wrapped by a context provider', () => {
       expect(subject()).toMatchSnapshot();
@@ -21,6 +29,19 @@ describe('EditorContext', () => {
       const wrapper = subject({ value });
 
       expect(wrapper).toHaveProp('value', expect.objectContaining(value));
+    });
+
+    it('calls getDraft and getPublished on mount', () => {
+      const getDraft = jest.fn();
+      const getPublished = jest.fn();
+
+      subject({
+        render: mount, // for useEffect
+        value: { getDraft, getPublished }
+      });
+
+      expect(getDraft).toHaveBeenCalledWith('test-template', '123');
+      expect(getPublished).toHaveBeenCalledWith('test-template', '123');
     });
   });
 });
