+++ src/pages/templates/tests/EditAndPreviewPage.test.js
@@ -1,33 +1,20 @@
 import React from 'react';
-import { mount, shallow } from 'enzyme';
-import useRouter from 'src/hooks/useRouter';
+import { shallow } from 'enzyme';
 import useEditorContext from '../hooks/useEditorContext';
 import EditAndPreviewPage from '../EditAndPreviewPage';
 
-jest.mock('src/hooks/useRouter');
-jest.mock('../components/EditPrimaryArea', () => { // todo, remove with use of mount()
-  const EditPrimaryArea = () => null; // must assign to variable for displayName
-  return EditPrimaryArea;
-});
-jest.mock('../components/EditSection', () => { // todo, remove with use of mount()
-  const EditSection = () => null; // must assign to variable for displayName
-  return EditSection;
-});
 jest.mock('../hooks/useEditorContext');
 
 describe('EditAndPreviewPage', () => {
-  const subject = ({ editorState, render = shallow } = {}) => {
+  const subject = ({ editorState } = {}) => {
     useEditorContext.mockReturnValue({
       draft: { id: 'test-template', name: 'Test Template' },
-      getDraft: () => {},
-      getPublished: () => {},
       hasDraftFailedToLoad: false,
       isDraftLoading: false,
       ...editorState
     });
-    useRouter.mockReturnValue({ requestParams: { id: 'test-template', subaccount: '123' }});
 
-    return render(<EditAndPreviewPage />);
+    return shallow(<EditAndPreviewPage />);
   };
 
   it('renders a page', () => {
@@ -42,18 +29,5 @@ describe('EditAndPreviewPage', () => {
   it('redirects and alerts when fails to load', () => {
     const wrapper = subject({ editorState: { hasDraftFailedToLoad: true }});
     expect(wrapper.find('RedirectAndAlert')).toExist();
-  });
-
-  it('requests draft and published templates on mount', () => {
-    const getDraft = jest.fn();
-    const getPublished = jest.fn();
-
-    subject({
-      editorState: { getDraft, getPublished },
-      render: mount // todo, remove when useEffect is supported by enzyme
-    });
-
-    expect(getDraft).toHaveBeenCalledWith('test-template', '123');
-    expect(getPublished).toHaveBeenCalledWith('test-template', '123');
   });
 });
